+++ test/parse/vhdl2008.vhd
@@ -66,4 +66,10 @@ begin
 
     /* Comments /* do not nest */
 
+    process is
+        variable x, y : integer;
+    begin
+        x := 1 when y > 2 else 5;       -- OK
+    end process;
+
 end architecture;