+++ test/test_simp.c
@@ -1097,6 +1097,33 @@ START_TEST(test_issue437)
 }
 END_TEST
 
+START_TEST(test_condvar)
+{
+   set_standard(STD_08);
+   input_from_file(TESTDIR "/simp/condvar.vhd");
+
+   tree_t a = parse_check_and_simplify(T_ENTITY, T_ARCH);
+
+   tree_t p = tree_stmt(a, 0);
+   fail_unless(tree_kind(p) == T_PROCESS);
+   fail_unless(tree_stmts(p) == 1);
+
+   tree_t s0 = tree_stmt(p, 0);
+   fail_unless(tree_kind(s0) == T_IF);
+   fail_unless(tree_conds(s0) == 2);
+
+   tree_t c0 = tree_cond(s0, 0);
+   fail_unless(tree_has_value(c0));
+   fail_unless(tree_stmts(c0) == 1);
+   fail_unless(tree_kind(tree_stmt(c0, 0)) == T_VAR_ASSIGN);
+
+   tree_t c1 = tree_cond(s0, 1);
+   fail_if(tree_has_value(c1));
+   fail_unless(tree_stmts(c1) == 1);
+   fail_unless(tree_kind(tree_stmt(c1, 0)) == T_VAR_ASSIGN);
+}
+END_TEST
+
 Suite *get_simp_tests(void)
 {
    Suite *s = suite_create("simplify");
@@ -1137,6 +1164,7 @@ Suite *get_simp_tests(void)
    tcase_add_test(tc_core, test_genmap);
    tcase_add_test(tc_core, test_issue436);
    tcase_add_test(tc_core, test_issue437);
+   tcase_add_test(tc_core, test_condvar);
    suite_add_tcase(s, tc_core);
 
    return s;